Implemented FlightPlan priority.
High priority flightplans will be executed before lower priority ones.
If a lower priority flightplan is currently executing commands to the drone, and a new higher priority flightplan has been ordered to be executed. The low priority one will be terminated, and whatever is left of it will be stored in the MessageManager to be executed later, the new higher priority FlightPlan will then be prepared and executed instantly.
Added sorting of FlightPlans, High priority -> low priority.

Changed the JSON datastructure (once again) to include a delay between DroneCommands.

Fixed an incapsulation inconsistency.

Updated JavaDoc and added more describing JavaDoc.